backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
Pillow==10.2.0
python-decouple==3.8
django-filter==23.5
dj-database-url==2.1.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
reportlab==4.0.9
django-import-export==3.3.5
openpyxl==3.1.2